+++ src/constants.js
@@ -21,6 +21,7 @@ export const DATA_EXPORT_API = 'data-export';
 export const DEPARTMENTS_API = 'departments';
 export const ELECTRONIC_ACCESS_RELATIONSHIPS_API = 'electronic-access-relationships';
 export const HOLDINGS_NOTE_TYPES_API = 'holdings-note-types';
+export const HOLDINGS_SOURCES_API = 'holdings-sources';
 export const HOLDINGS_TYPES_API = 'holdings-types';
 export const IDENTIFIERS_TYPES_API = 'identifier-types';
 export const INSTANCE_FORMATS_API = 'instance-formats';