+++ src/routes/ConsortiumManager/settings/inventory/holdings/HoldingsSources/index.js
@@ -1,0 +1,1 @@
+export { HoldingsSources } from './HoldingsSources';